+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `config.py` for migration
 - Added parser for `gambitcore`
 - Added tests and test files for `gambitcore`
+- Added `sccmec` parsing, models and tests
 
 ### Fixed